requirements.txt
isort
pandas
mysql-connector-python
python-dotenv
loguru
boto3
fsspec
s3fs
gspread
oauth2client
google-api-python-client
scikit-learn
PyYAML